--- topothehourbot/channels/__init__.py
@@ -1,2 +1,0 @@
-from .ircv3 import *
-from .sql import *
