--- topothehourbot/channels/ircv3/__init__.py
@@ -1,2 +1,0 @@
-from .protocols import *
-from .twitch import *
